Clarify and speed up HeapSnapshot.isEssentialEdge
The functions buildPostOrderIndex and buildDominatorTree both operate on
a subset of the graph, ignoring edges which should be treated as weak
based on several rules. This change centralizes those rules in a single
function to improve readability in buildPostOrderIndex and
buildDominatorTree. This change also adds a bitmap to cache the results,
which provides a substantial speedup on large graphs: on my current
favorite test case, the time in HeapSnapshot.initialize drops from 59
seconds to 46 seconds.

Bug: 42203857
Change-Id: Ifa5512d0ebe9786f9b3ef0c6086f39becd75bfad