[esbuild] Use correct tsconfig.json to build
ESBuild used the root tsconfig.json which is only intended for
VSCode. This still mostly works as we have a "base" tsconfig.json that
acts as the source of truth for the different compiler options
globally.

This CL updates ESBuild to use the per ts_library invocation
generatd tsconfig.json instead.

R=pfaffe@chromium.org

Fixed: 352249693
Change-Id: I6936f66b2e2f78aa7971dc2158076d697dc15553