[Freestyler] Fix waiting for freestyler binding call
I have noticed that I forgot to return the promise
from freestyler calls. This resulted in Freestyler
getting a response before the styles were applied.

Unfortunately, proper promise handling revealed another
limitation that the console APIs are only available
in the initial scope and are gone from the scope
in the next microtask.

This CL works around the issue by re-mapping console APIs
to local vars and using IIFEs to wrap LLM-generated
scripts.

Bug: 360776969
Change-Id: If82c616953ce05400d6619db84108c69c6c99681