Add DataGridWithPreview class
This renders a table of key-value pairs in a DataGrid with a
corresponding preview widget. This will be helpful in upcoming work to
show extension storage in DevTools.

As an example, this CL uses the new helper class to greatly simplify the
implementation of DOMStorageItemsView.

Some use cases (e.g cookies) require more than a simple key-value
schema. That is not supported here to reduce complexity in the CL but
should be easy to add support for in a future CL.

Bug: N/A
Change-Id: If332b397b2fa92d2c8b2cacb95619883625aac29