RPP: fix overlay visibility
We had a bug where sometimes some overlays would incorrectly stay
visible even though they should be hidden.

The problem was we were controlling this via the `visibility` property,
but an element with `visibility: hidden` may still have its children
visible if they have their visibility set to `visible`. This was the
root of the bug; we were setting the parent overlay to hidden, but that
did not apply to all children.

This CL moves us to `display: block` and `display: none` which does not
have this problem.

Fixed: 356352905,356359407,356353052
Change-Id: I8479a237e9f47bbbf2f89789367a8409f0303f28