RPP: remember user state for sidebar
It was annoying that on each trace import the sidebar would
automatically open. This CL updates the sidebar state to persist, so if
the user last had it closed, we won't automatically pop it open.

The only exception is the very first time the user imports a trace after
we ship the sidebar; in this case we make it appear to aid with
discovery. But once this has happened once we persist this setting to a
global storage variable, so it will not happen again.

Fixed:360096750
Change-Id: I2561fdcb956d05a37023a61f91ae31a95851c15f