Fix interactions tsconfig
In crrev.com/c/5629076 as a drive-by I included the root tsconfig in the
interactions folder, but that was using a hammer to crack a nut and it
caused some unexpected TypeScript errors in the interactions tests.

The issue comes because unexpectedly we want to import from the
front_end folder in the interaction tests - not normally done - and so
you get a bunch of warnings in your editor (it does not impact the
actual build).

Extending the root tsconfig seems to cause confusion with TS and how it
resolves some of the Puppeteer types so this CL localises the solution
into just the performance/timeline folder and tells TS that it is OK to
include files from the root front_end and test dir within there.

Bug: none
Change-Id: I3e923b1b9c874784ee3dd5bdfc9649abec696c79