fix: update the icon state of the expanded element after editing
When a CSS property is expanded, double-clicking on the property value to enter edit mode will temporarily collapse the expanded content. After exiting edit mode, the content will return to its previously expanded state. However, because the expand button is newly added, it defaults to a collapsed state, which is inconsistent with the current actual state.

Bug: https://issues.chromium.org/issues/362887150
Change-Id: Ia14b0edf9765327e948e63c14befac91c7f918f9