[RPP] Include top culprits to CLS insight
This shows the first 3 culprit types based on the worst clusters. This
also includes clusters in the insight return, which will be helpful for
the overlay down the line.

(in this example, it was just a bunch of font reqs)
https://screenshot.googleplex.com/6fwZfkj9wBwKhR3

TODO: adding a reference to the cluster event details. We don't have
this event reference in the UI yet.

Bug:329735547
Change-Id: I1812e51bddd604f0ffaf590f389f7a3c1304675a