[RPP]  When calculate the initiators, remove the loop to avoid crash
The main reason to not doing this in the TraceEngine is to reduce the time when loading a trace file, and defer it to when users click an entry.
Also usually there will be dozens ~ hundred event_to_initiator pairs and for a single event usually less than 5 initiators chain, so it will be much more efficient to remove the loop after user clicking.

Fixed: 363898657
Change-Id: If8c9dac528ef9cd8d4a5088126835e012f18aa11