[RPP] Navigate Breadcrumbs Instead of removing them
We are currently working on features that will need to automatically navigate to a window in the timeline and, if a window we need to navigate to is outside of the created breadcrumb, the only way to navigate to that window would be to destroy breadcrumbs.

To avoid destroying breadcrumbs and still being able to navigate to any window in the timeline, this cl adds an ability to navigate breadcrumbs without removing them. Now clicked breadcrumb becomes active without deleting the following breadcrumbs.
Child breadcrumbs can be overwritten by creating a new breadcrumb or removed through a context menu.

Video: http://screencast/cast/NTcyMDkzMDM3MzMzNzA4OHw3OTUxOWZkOS0zNQ

Bug: 362726217
Change-Id: I92d077731468c7a2286caffa4b804811dbd12f47