[scopes] Add SourceMapScopeInfo.hasVariablesAndBindings()
Source map generators may choose to emit scopes that only carry
positions and names, but not variable names and binding expressions.
Such a source map is still valuable since it allows translating
stack traces more reliable than with just mappings. But we can't
use it to show a scope view.

This CL adds a simple heuristic check to see if variable names and
binding expressions are present.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: I9480ddeee322ef981f81063d788e18f11d0dacf4