[sdk] Add central method to report source map problems
This CL adds a new 'reportSourceMapProblem' method to the SourceMap
class. For now it logs the warnings/errors to the user's console panel.
In the future we might also route it to the "Developer resources"
panel (in addition or instead).

THe main target are tooling developers trying to implement new
source map features like the "Scopes" proposal and need easy feedback.

Follow-up CLs will instrument more places where source map processing
or usage can fail. At the moment we emit an error for all decoding
errors.

R=kimanh@chromium.org

Bug: 40277685
Change-Id: Ic36cdec348c8dea7095181c8106012550cdc4474