[scopes] Resolve original scope chains using source map 'scopes' info
There are 2 ways how we can construct a original scope chain based on a
V8 paused position:

  1) We find the inner-most generated range that contains the V8 paused
     position. We'll walk the generated range chain outwards until we
     find a range that also represents an original scope. That original
     scope is than the inner-most original scope and we walk that
     original scope's parent chain outwards to construct the original
     scope chain.

  2) We use the source map's mappings to map the V8 paused position
     back to an original position. With the original paused position
     we can then search for the inner-most original scope. The
     byproduct of that search is the result, the original
     scope chain.

1) and 2) can produce different results when the mapped paused
position lies in a different original scope than the inner-most
generated range's original scope (e.g. minifiers that constant fold
or otherwise aggressively inline expressions).

We use approach 2). This is because we have to stay consistent with
the rest of the DevTools UI. Otherwise the obtained scope chain would
not fit with the paused position in the editor view.

The CL also adds a couple of unit tests to document the existing
behavior and design choices. These might need to be revisited
depending on how source map generators actually choose to emit
scopes.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: Ia033d3a5f8d1c60be5059c05da770e220940538e