[scopes] Fix scope chain construction for inlined functions
This CL builds on https://crrev.com/c/5833561 and fixes the original
scope chain when dealing with inlined functions.

The DebuggerModel currently inserts "virtual" call frames for
inlined functions. Virtual call frames have an inlineFrameIndex > 0.

When we build a scope view, it's not enough to find the inner-most
generated range. We then need to check the inlineFrameIndex and
drop inlined ranges. This is very similar to unwinding a stack but
limited to inlined frames.

We might change this in the future depending if we still want to
handle inline frame expansion on the DebuggerModel level or if
we push it to an higher abstraction level.

R=kimanh@chromium.org

Bug: 40277685
Change-Id: I02e86e77f91462a36a512309e8097a38da2dabb9