[RPP] Construct tracingCompletePromise well before it is needed
It seems feasible that 'tracingComplete' could be invoked before
'stopRecording', which would cause 'waitForTracingToStop' to never
resolve. I was not able to reproduce the issue on purpose, so hard
to verify as a fix.

This refactor create a promise just before starting the trace, so now
it is obviously always in a valid state.

This somehow also fixes a recent regression where loading the trace
took an additional 5s (see 364909878).

Bug: 364921460, 364909878
Change-Id: Ie293eeaf44d1abfe43fe9c870edecbc5bef21f2a