[scopes] Improve type for inlining info
This CL changes the return type of 'findInlinedFunctions' to a
dedicated 'InlineInfo' type. It clears up the distinction between the
original function name of the surrounding function vs the names of
all the functions that were actually inlined.

The CL also simplifies the logic a bit, since the check for the
original scope's kind is not truly necessary. We assume that the
presence of 'callsite' is marker enough.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: Iea10f3f10ac688fc3f78c124d83362d265baa225