[scopes] Move 'expandCallFrame' into SourceMapScopesInfo
This CL moves 'expandCallFrame' out of SourceMap ino
SourceMapScopesInfo and adds a bunch of unit test to get more coverage.

Drive-by: This CL also fixes the CallFrame class: creating virtual
call frames now also copies the exception object if present, and if
a function name is passed via constructor it's preferred over the
payload one even in the case where we pass an empty function name.

The current architecture is less than ideal: If "scopes" proposal
scopes are present, we translate the function name on the `CallFrame`
directly, but if we use source map 'mappings' plus 'names', we
resolve it later from the UI via 'NamesResolver'. Mid-term we'll
clean this up by distinguishing "raw" stack traces received by the
engine, and UI-ready stack traces that were augmented with debug
info.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: I1ffd8a2f98e157b1b735b99f7b43b06a7d7a58f3