RPP: show initiators for non-script network requests
The logic to use the stack trace works for requests that have a stack
trace, but for other requests we can still try to match based on the
URL and show something to the user. Especially as they see the arrow in
the flame chart, it's confusing if they then don't see anything in the
details.

Fixed: 363242792
Change-Id: I0a637ed730b0f34e732d929a5c757799bf130e11