RPP: move EventUICategory into components/EntryStyles
This is needed so that we can use the styles in the code that calculates
the user visible title for a given event, as lots of events have custom
names as defined in the styles map.

This CL doesn't change any implementation, it's just the mechanical
move to set me up for future CLs.

Bug: 365047728
Change-Id: I9cbaface34883c104b262ec6735b6f1c8a0eede0