[cleanup] Remove redundant 'Save all as HAR with content' menu item.
This option does exactly the same as the Download icon in the toolbar,
so it's 100% redundant. Given that the title is also different from the
tooltip of the Download icon, it's probably even misleading to
developers and they might wonder if the context menu item does something
different. In either case, we concluded as part of the design document
mentioned below that we don't want to have this redundant menu item.

Doc: https://goo.gle/devtools-har-exclude-cookies-design
Bug: 361717594
Change-Id: I7e7311b7e7779e5df6f9cff2d88334a66d35d57a