[network] Export sanitized HARs by default.
This CL introduces a new disabled-by-default setting "Allow to generate
HAR (with sensitive data)", which guards the availability of options to
export/copy HARs with Cookie, Set-Cookie, and Authorization headers. By
default the exported/copied HARs will be sanitized to not include cookie
or authorization information.

This is reflected by renaming the titles of the actions accordingly to
"Copy all as HAR (sanitized)" and "Export HAR (sanitized)". Only when
the new setting is explicitly enabled by the user, a second set of
options becomes available, which allows to "Copy all as HAR (with
sensitive data)" and "Export HAR (with sensitive data)".

Please refer to the design document for further details and screenshots.

Doc: https://goo.gle/devtools-har-exclude-cookies-design
Bug: 361717594
Change-Id: I989175eec061c2f58bbd734a2f9e58a7b727bdba