Take the setting's disableCondition into account before generating a console insight
In the new flow, the hover button is shown despite the feature being
blocked by locale, age, or geographic location. This CL ensures that
no insights are being generated in these cases. Instead the user sees
a link to settings, where they can see the reason for the feature not
being available.

Drive-by: The message which links to settings needs to be localized,
since it can be seen by users with a non-English locale.

Fixed: 365706617
Bug: b/350668580
Change-Id: I2ddb9b80f66c6362591dd45b36db68f73e18c5c8