RPP: hide entry links if either is in a collapsed track
This CL adds the ability to calculate a group for a given entry, and
therefore determine if the entry is in a collapsed group or not. If we
have an ENTRIES_LINK overlay that has at least one event in a collapsed
track, we hide it.

Note that there is more work to do here; if the user clicks to reveal an
annotation that is in a collapsed group, we want to expand the group. I
will do this in a follow-up CL.

Bug: 365493598
Change-Id: Ifcb9b6956159d5d50051b49fa198e231739a7793