[RPP] Navigate to the smallest breadcrumb that includes a window
It is possible to navigate to different overlays from the sidebar, however the overlay could be outside of the current active breadcrumb.

This CL adds a method to the Breadcrumbs class that navigates to the smallest breadcrumb that includes a window we need to display (only if the current active breadcrumb window does not include that window).
This method is called before revealing overlays selected from the the sidebar.

Video: http://screencast/cast/NTcwNTEyMjAwODAwNjY1NnwxOTZhZWU5Yi0xZQ

Bug: 362726217
Change-Id: Ie9a6188614e2d713ad597949caecc9636c1986a2