Adding Source Map Server extension APIs (1/2)
This PR adds the APIs required for Source Map Server implementation.
The developer will register a plugin that implements the
"SourceMapServerExtensionPlugin" interface. This plugin will get
notified everytime a JS script is loaded and the extension developer
will be able to respond to this notifications (via a Promise) with a
sourcemap that fits the script.

This first part only covers APIs shape and implementation, wiring
for notification from Debugger SDK/Models, along with
source map passing is pending for part 2.

Full spec can be found here (Implementation):
https://docs.google.com/document/d/1x9_MTMVr213ApruMK6ryc4JpPHahvOBlExP6sHkMv8Y/edit?pli=1

Bug: 364508694
Change-Id: I05d379eaaf10ddd0bc53133d3b4000585aeadd53