Modify millisToString to always render 0 with a unit.
When higherResolution is true, this renders "0 μs". Otherwise it renders
"0 ms". Always presenting the user with the intended resolution seems
like a win for any usage of this function.

This change was encouraged by a table rendering numbers with
millisToString, where numbers are often 0 or near 0. It was very odd to
see "0" and "0 ms" in the same column.

Bug: 366290121
Change-Id: I726077fd09ab2a4c902cb0cc328aec815ee984bc