WIP: Desaturate flamechart to highlight search results
To demo: invoke find (ctrl-f) on a trace and type in 'script' or w/e.

And SVG with mix-blend-mode desaturates the content visually beneath it,
and the SVG has a mask with cutouts within it allowing portions to
'punch through' and not be affected by the desaturation. Shades of grey
within the SVG mask also allow partial saturation.

The SVG layer can fall out of sync with the canvas. There remains research
to determine if that can be resolved or if a canvas-based solution is preferred.

Bug: none
Change-Id: I48f4628f07d26feddb2b8e1df580077593b9766e