[sources] Don't reset the current search index when focusing the editor.
Previously there was a race condition between setting the current search
index after clicking the next/prev arrows and clearing the current
search index due to focusing the CodeMirror editor. The dispatching
of the scroll position changes in CodeMirror can lead to 'focus'
events being emitted, which is the cause for the race.

The original motivation for resetting the current search index inside
`onFocus()` was likely to get a clean state after switching panels, but
that isn't necessary, since switching back and forth between Sources and
some other panel doesn't invalidate the search results, and therefore
the index is also still correct.

Fixed: 361600646
Change-Id: If76edafc84950e70368867c4d8b7a70d67f0effc