RPP: fix time range labels being editable immediately
This CL fixes a bug where a trace with loaded time range labels would
all be editable immediately because we didn't set their edit mode to
false when constructing them from a file.

Additionally this CL also fixes a bug I found where the overlays render
before the flamechart is fully complete. This causes the canvas rect
that is cached in FlameChart.ts to be a rect with a width of 0, which is
obviously incorrect and causes the calculations to be messed up. Now we
used the cached rect only if it has a width and height, and if it
doesn't then we recalculate it.

As a drive-by improvement I also added a `bulkAddOverlays` method that
adds all the overlays before calling `update()` which is a little more
efficient then adding + updating for each one.

Bug: none
Change-Id: Id15c2d3f737b75ff56fc57f3d33b8b1c357570c6