DevTools: Add ExtensionStorageModel and initial UI for storage viewer
Adds an ExtensionStorageModel which creates new ExtensionStorage
instances when an execution context associated with an extension is
created.

This is added behind a new experiment intended to exist while this feature
is worked on over a few further CLs.

Currently each storage area shows as a new top-level entry. In a future
CL we can update this to provide a tree-based layout.

Additionally, there is no logic to hide storage areas if the storage API
is not enabled for an extension. This will also be added as a follow-up.

Bug: 40963428
Change-Id: Ic0c532b4ae47c1a1e55fcbbff5a46708285f8648