[RPP] Add overlay effect on hover for Third Parties insight
This adds a new mechanism for insights to provide overlays: overlay
overrides.

Insights can call `this.onOverlayOverride(overlays)` with a list of
overlays to show _instead_ of the ones created by `createOverlays`. To
restore the original overlays, call the same method with `undefined`.

This could be extended later as use cases require (for example, to
augment the overlays rather than replace them).

Drive-by change: added `this.bulkRemoveOverlays` and used
`this.bulkAddOverlays` more to reduce number of calls to
`#overlays.update()` (though there are still two per update, needs more
refactoring)

Bug: 339795725, 352244718
Change-Id: I4a7b7cef77731db8231097c4bd4ebbb5f4996b97