[RPP] Add overlay effect on hover/click for Third Parties insight
Hovering/clicking on the table rows in Third Parties now shows overlays
for just those origins. Unhovering/unselecting reverts the overlays back
to the original set created by the insight.

To support this, a new mechanism is added for insght overlays: overlay
overrides.

Insights can call `this.onOverlayOverride(overlays)` with a list of
overlays to show _instead_ of the ones created by `createOverlays`. To
restore the original overlays, call the same method with `null`.

This could be extended later as use cases require (for example, to
augment the overlays rather than replace them).

Drive-by change: added `this.bulkRemoveOverlays` and used
`this.bulkAddOverlays` more to reduce number of calls to
`#overlays.update()` (though there are still two per update, needs more
refactoring)

Bug: 339795725, 352244718
Change-Id: I4a7b7cef77731db8231097c4bd4ebbb5f4996b97