[RPP Observations] Elide node link text
This solves the focus box problem by eliding the link text instead of
using `overflow: hidden`. Visually eliding is better as well.

This did require a modification to all link node styles such that the
maximum width of the button element can be controlled by the component
using the link node.

Bug: 367885814
Change-Id: Ib99a6794439c634d644564387f7a3e8d5e5e0d7f