Show panel sidebar icons
Icons were being created in this function, but then immediately
replaced by the new title. I assume this was unintentional. The fix
is to only replace the last child since this element always adds the
title as the last element. Rather than replacing all children with
the title, only replace that old title.

This also fixed the issue where the 'Overview' button would not
light up when hovered over. See below for changes:

https://screenshot.googleplex.com/BugBiXH2MFixxdJ.png


Bug: 368603997
Change-Id: Ibbce54ff9da746c96bb198bf76db145389a4c59f