Memory: Use the Lengauer-Tarjan algorithm to compute retained sizes
The current algorithm used to build the dominator tree is performant
only for graphs with upto 30,000 nodes as per the original paper[1].
Even the smallest heap snapshot (of about:blank) has around 30,000 nodes
and hence this algorithm scales poorly to larger, more complex sites.

The Lengauer-Tarjan algorithm[2] is the best-known and most widely used
of the fast algorithms for dominance[1]. The "simple" version of the
algorithm (implemented in this CL) has a timebound of O(E log (N)) as
opposed to O(N ^ 2) for the current algorithm and hence scales much
better to sites with millions of nodes while still being marginally
faster for smaller sites, as the numbers below show[3]:

| Website             | Heap size | Node count | Old algo | New algo |
|---------------------|-----------|------------|----------|----------|
| about:blank         |    1.2 MB |     26,990 |    31 ms |    23 ms |
| wikipedia.org       |    7.7 MB |    133,529 |   114 ms |    65 ms |
| amazon.com          |   10.0 MB |    237,114 |   190 ms |   121 ms |
| google.com          |   18.5 MB |    459,426 |   357 ms |   263 ms |
| youtube.com         |   66.3 MB |  1,537,407 |  2900 ms |  1026 ms |
| teams.microsoft.com |  157.0 MB |  3,793,247 | 20573 ms |  3111 ms |

# Implementation notes
Appendix B in the paper[2] lists the complete algorithm whose "simple"
version is implemented as-is in this CL, with the following tweaks:
- Iterative versions of dfs() and compress() are used since recursion
  hits stack limits.
- Heap snapshots often contain unreachabled nodes. These are visited
  separately after the initial DFS from the root.
- Only essential edges are considered while traversing the graph.
- The dominator of the root is set to the root itself instead of 0.

This CL also updates the computeIsEssentialEdge() function to treat self
edges as non-essential since they don't retain the object itself.

# Review notes
- Patchset 10 has the final logic of the algorithm and includes code to:
    - Measure the runtimes of the two algorithms, and
    - Verify that the outputs of the two algorithms match.
- Patchset 11 removes all the validation and measurement code and has
  the final version of the new function, including code comments.
- The old functions are fully removed in Patchset 12.

# Footnotes
[1] https://www.clear.rice.edu/comp512/Lectures/Papers/TR06-33870-Dom.pdf
[2] https://doi.org/10.1145/357062.357071
[3] These measurements were done on Chrome Version 129.0.6668.59 using a
    release build of instrumented DevTools on a Windows x64 machine with
    16 cores and 64 GB memory. 5 heap snapshots were taken for each site
    and the numbers from the final snapshot are reported here. The
    measured time includes graph traversal and computation of retained
    sizes for both algorithms. Patchset 10 is the exact code run.

Bug: 363034702
Change-Id: I8195d67ef6ff0406736a360f1f57f154859aa354