RPP: allow the context menu to open with a keypress (space)
This CL adds support for the user pressing space when an entry is
selected to trigger the context menu at this point. This is how we can
make the creation of annotations keyboard accessible as once the context
menu is open it can then be navigated with the arrow keys.

Bug: 369201266
Change-Id: I6ee13ce5a77f93c6666e4f0874d54950af59c141