[RPP] Store code location source mappings
Just as we are doing with function names, this CL adds support stores
existing code location mappings (i.e. script url, line and column
numbers) whenever we detect that a source map has been attached in
devtools.

With this in place, we can more accurately represent source locations in
all places in terms of the authored code. For example, when evaluating
if an entry's location is in the user ignore list we can first resolve
the corresponding authored code and then determine if it's ignore
listed (which this CL also does). This is consistent with what the
debugger does and with what devs expect in general.

In a follow up, I'll use this new functionality in more places in the
Performance panel.

Fixed: 358572661
Change-Id: I232f83bf0de22c244e14f09e7c3013731070f8d6