RPP: make entry links creatable with a keyboard
This CL makes entry links creatable with the keyboard, building on the
fact that ContextMenus are now keyboard-openable.

As a side I also found a bunch of confusion (in my head!) around how
entries are selected. We have different events and it is really not very
clear. I left some comments to try to clarify this!

Bug: 369201266
Change-Id: I678c7f9ed20214e305f9fc32e6bd277e1037e48f