[RPP] Refactor Insights Table component
Moves the logic for responding to user interaction to be fully
encapsulated in the Table component, rather than requiring the parent
component to do it. We can do this now that we have decided on a
consistent hover UX for tables within insights. We don't need
the previous level of customization.

To support this, Table now has a TableRow type (rather than being an
array of strings), so the component itself knows the relevant `overlays
to display when the user interacts the row.

No user-facing changes yet.

Bug: 369102516
Change-Id: Ia2a06ecc1dcb5fc61d0957c4961a29ee32416012