Refactor the Security panel sidebar
This changes the security panel's sidebar to make it more modular,
which will help when adding privacy tools in the future. The code
pattern was taken from the application panel's sidebar. This allows
each new tree element to have it's own class and be added into the
three in standard manner.

Side by side comparison (refactored left, original right): 
https://screenshot.googleplex.com/47PTmo4kN8rd7Cd.png

Bug: 365737493
Change-Id: I904774ae2b4d7549d254b6b063aca96182531539