[sources] Fix broken replace toggle in searchable views.
This broke as part of https://crrev.com/c/5563488, which switches the
toolbar buttons to use the button component. As part of this the logic
to toggle upon clicks is already built into the toolbar button now,
and the additional code to flip the toggle in the "click" handler
caused the toggle to immediately go back to the previous state.

Fixed: 369781726
Bug: 325441197
Change-Id: Ic02fa557d512db383ce8f903d22212ca2e34cbbf