[PP] Switch appending order of user timings on equal start and end
User timing entries come as trace events dispatched when
performance.measure/mark is called. The trace events buffered in
devtools frontend are sorted by the start time. If their start time
is the same, then the event for the first call will appear first.

When entries are meant to be stacked, the corresponding
performance.measure calls usually are done in bottom-up direction:
calls for children first and for parent later (because the call
is usually done when the measured task is over). This means that
when two user timing events have the start and end time, usually the
second event is the parent of the first. Hence we should switch the
event order in case of equal start and end times, because we should
append parent tasks first so that they appear on top.

Bug: 358307050
Change-Id: Ic20a609f231724c63136ce713c8e5b868c3cbce3